feat(cli): adding compile schemas to maker cli
