+++ Makefile
@@ -16,7 +16,7 @@ public/schemas/%.schema.json: src/schemas/%.schema.json
 
 src/schemas/tiles.defs.json: src/schemas/fields.schema.json src/schemas/tiles.src.json
 	@echo "Compiling $@"
-	@node ./bin/compileSchemas.cjs
+	@node ./bin/maker.js compile
 
 clean:
 	@echo "Removing output folders"
